+++ webapp/src/main/java/ru/tele2/govorova/java/pro/webapp/LoggerFilter.java
@@ -1,15 +1,9 @@
-package web;
+package ru.tele2.govorova.java.pro.webapp;
 
+import jakarta.servlet.*;
+import jakarta.servlet.annotation.WebFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import jakarta.servlet.Filter;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.FilterConfig;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.annotation.WebFilter;
 
 import java.io.IOException;
